+++ RELEASE-NOTES.md
@@ -67,6 +67,9 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Inlet: Cleaned up `Table` interface to eliminate ambiguity and duplicated functionality
 - Inlet: Renamed `DocWriter` to `Writer` and refactored its interface
 - Inlet: Renamed `Table` to `Container`
+- Inlet collections of mixed or incorrect type will now fail verification, even if they're
+  not marked as required
+- Required collections no longer fail Inlet verification if they are empty in the input file
 
 ### Fixed
 - Updated to new BLT version that does not fail when ClangFormat returns an empty